// src/Play/Connection/ReceivedInput/Types/CharacterStatTypes.ts
export interface ICharacterHealth {
    characterId: string;
    health: number;
}

export interface ICharacterMana {
    characterId: string;
    Mana: number;
}

export interface ICharacterStats {
    characterId: string;
    maxhealth: number;
    maxMana: number;
}

export interface ICharacterLevel {
    characterId: string;
    level: number;
}